refactor(mb-api): fold panierTauxProgression module into panier
Suite review PR #2210 : un module par domaine côté API. La query et le resolver
de calcul du taux agrégé d'un panier vivent dans panier/, et la route est servie
par panierRoutes. Suppression du path tsconfig dédié.
